requirements.txt
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.7
django-bootstrap-icons==0.7.3
django-bootstrap5==21.1
django-import-export==2.6.1
django-stdimage==5.3.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
lxml==4.6.4
MarkupPy==1.14
numpy==1.21.4
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.4
patsy==0.5.2
Pillow==8.3.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
static3==0.7.0
statsmodels==0.13.1
tablib==3.1.0
toml==0.10.2
urllib3==1.26.7
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
